backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pandas==2.1.0
numpy==1.24.3
beautifulsoup4==4.12.2
requests==2.31.0
scikit-learn==1.3.0
google-generativeai==0.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
openpyxl==3.1.2
torch==2.0.1
huggingface_hub==0.20.0
transformers==4.36.0
